requirements.txt
# Strategic Career Platform v2.0 Requirements
# Python 3.9 or higher required

# Core Dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0

# Database (SQLite is built-in to Python)
# sqlite3 is included with Python

# Web Scraping & Automation
beautifulsoup4>=4.12.0
selenium>=4.15.0
chromedriver-autoinstaller>=0.6.0

# Data Processing
pandas>=2.0.0
python-dateutil>=2.8.2

# Email Handling
email-validator>=2.0.0

# Google APIs (for Gmail integration)
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Document Generation
reportlab>=4.0.0
python-docx>=0.8.11
markdown>=3.4.0
jinja2>=3.1.0

# Optional: AI/ML Integration (uncomment if using)
# openai>=1.0.0
# anthropic>=0.5.0
# ollama>=0.1.0

# Optional: Development Tools
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0

# Optional: Advanced Features
# sentence-transformers>=2.2.0  # For semantic matching
# torch>=2.0.0                   # For ML models
# scikit-learn>=1.3.0           # For ML algorithms